hf_deployment/requirements.txt
# Core dependencies for Hugging Face Spaces deployment
streamlit>=1.28.0
pandas>=2.0.0
plotly>=5.15.0
numpy>=1.24.0

# AI and ML dependencies
groq>=0.4.0
autogen-agentchat>=0.6.0
autogen-core>=0.6.0
autogen-ext>=0.6.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4
huggingface-hub>=0.17.0

# Database (SQLite for demo, PostgreSQL for production)
sqlite3
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0
pydantic>=2.0.0

# Optional dependencies for enhanced features
openpyxl>=3.1.0
Pillow>=10.0.0
python-dateutil>=2.8.0

# For Hugging Face Spaces compatibility
gradio>=3.40.0
